Add REACTER files builder and modifiers
